requirements.txt
django==5.1.7
djangorestframework==3.16.0
drf-yasg==1.21.10